Move SpriteDisplay component to kas::layout
