+++ crates/kas-core/src/layout/size_types.rs
@@ -5,10 +5,10 @@
 
 //! Types used by size rules
 
-use super::SizeRules;
-use crate::cast::{Cast, Conv, ConvFloat};
+use super::{AlignHints, AxisInfo, SizeRules};
+use crate::cast::{Cast, CastFloat, Conv, ConvFloat};
 use crate::dir::Directional;
-use crate::geom::Size;
+use crate::geom::{Rect, Size, Vec2};
 
 // for doc use
 #[allow(unused)]
@@ -147,6 +147,112 @@ pub enum Stretch {
 impl Default for Stretch {
     fn default() -> Self {
         Stretch::None
+    }
+}
+
+/// Scaling of image according to scale factor
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash)]
+pub enum SpriteScaling {
+    /// Do not scale with scale factor
+    Original,
+    /// Use the nearest integer of scale factor (e.g. 1, 2, 3)
+    Integer,
+    /// Use raw scale factor
+    Real,
+}
+
+impl Default for SpriteScaling {
+    fn default() -> Self {
+        SpriteScaling::Integer
+    }
+}
+
+/// Scaling of image sprite within allocation
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash)]
+pub enum AspectScaling {
+    /// Align sprite within available space without further scaling
+    None,
+    /// Scale sprite to available space with fixed aspect ratio
+    Fixed,
+    /// Scale sprite freely
+    Free,
+    // TODO: we could add repeat (tile) and mirrored repeat modes here
+}
+
+impl Default for AspectScaling {
+    fn default() -> Self {
+        AspectScaling::Fixed
+    }
+}
+
+/// Widget component for displaying a sprite
+#[derive(Clone, Debug, PartialEq)]
+pub struct SpriteDisplay {
+    /// Margins
+    pub margins: MarginSelector,
+    /// The native size of the sprite
+    pub size: Size,
+    /// Sprite scaling according to scale factor
+    pub scaling: SpriteScaling,
+    /// Sprite scaling within allocation, after impact of scale factor
+    ///
+    /// Note: this only has an impact if `stretch > Stretch::None`.
+    pub aspect: AspectScaling,
+    /// Widget stretchiness
+    pub stretch: Stretch,
+}
+
+impl Default for SpriteDisplay {
+    fn default() -> Self {
+        SpriteDisplay {
+            margins: MarginSelector::Outer,
+            size: Size::ZERO,
+            scaling: SpriteScaling::Integer,
+            aspect: AspectScaling::Fixed,
+            stretch: Stretch::None,
+        }
+    }
+}
+
+impl SpriteDisplay {
+    /// Generates `size_rules` based on size
+    ///
+    /// Set [`Self::size`] before calling this.
+    pub fn size_rules(&mut self, sh: &mut dyn SizeHandle, axis: AxisInfo) -> SizeRules {
+        let margins = self.margins.select(sh).extract(axis);
+        let size = self.size.extract(axis);
+        let size = match self.scaling {
+            SpriteScaling::Original => size,
+            SpriteScaling::Integer => i32::conv_nearest(sh.scale_factor()) * size,
+            SpriteScaling::Real => (sh.scale_factor() * f32::conv(size)).cast_nearest(),
+        };
+        SizeRules::new(size, size, margins, self.stretch)
+    }
+
+    /// Aligns `rect` according to stretch policy
+    ///
+    /// Assign the result to `self.core_data_mut().rect`.
+    pub fn align_rect(&mut self, rect: Rect, align: AlignHints) -> Rect {
+        let ideal = match self.aspect {
+            AspectScaling::None => self.size,
+            AspectScaling::Fixed => {
+                let size = Vec2::from(self.size);
+                let ratio = Vec2::from(rect.size) / size;
+                // Use smaller ratio, which must be finite
+                if ratio.0 < ratio.1 {
+                    Size(rect.size.0, i32::conv_nearest(ratio.0 * size.1))
+                } else if ratio.1 < ratio.0 {
+                    Size(i32::conv_nearest(ratio.1 * size.0), rect.size.1)
+                } else {
+                    // Non-finite ratio implies size is zero on at least one axis
+                    rect.size
+                }
+            }
+            AspectScaling::Free => rect.size,
+        };
+        align
+            .complete(Default::default(), Default::default())
+            .aligned_rect(ideal, rect)
     }
 }
