+++ crates/kas-widgets/src/lib.rs
@@ -101,6 +101,6 @@ pub use scrollbar::{ScrollBar, ScrollBarRegion, ScrollBars, Scrollable};
 pub use separator::Separator;
 pub use slider::{Slider, SliderType};
 pub use splitter::*;
-pub use sprite::{Image, SpriteDisplay, SpriteScaling};
+pub use sprite::Image;
 pub use stack::{BoxStack, RefStack, Stack};
 pub use window::Window;
